requirements.txt
python-dotenv
simplematrixbotlib
matrix-nio[e2e]
